Rename `user_id` field to `rightHolder_id` for search
We really want to make it look like a right holder is its own separate thing, even if currently it is the same. The idea is that a user later on will be able to manage multiple right holders, which may correspond to different artistic entities that are the same person.